[android][expo-media-library] fix broken native tests

+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/AssetsOptions.kt
@@ -4,11 +4,11 @@ import expo.modules.kotlin.records.Field
 import expo.modules.kotlin.records.Record
 
 data class AssetsOptions(
-  @Field val first: Int,
+  @Field val first: Double,
   @Field val after: String?,
   @Field val album: String?,
   @Field val sortBy: List<String>,
   @Field val mediaType: List<String>,
-  @Field val createdAfter: Int?,
-  @Field val createdBefore: Int?
+  @Field val createdAfter: Double?,
+  @Field val createdBefore: Double?
 ) : Record
