+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/assets/GetAssets.kt
@@ -37,7 +37,7 @@ internal class GetAssets(
           contentResolver,
           assetsCursor,
           assetsInfo,
-          limit,
+          limit.toInt(),
           offset,
           false
         )
